Rename indexes module to index and any references in the project - Update imports and docstrings for index in modules and tests - Change the Sphinx master doc name in the conf.py from index to default - Rename Sphinx doc indexes.rst to index.rst
