+++ docs/default.rst
@@ -1,0 +1,14 @@
+python-cloudant documentation
+=============================
+
+This is the official Cloudant client library for Python.
+
+.. toctree::
+   :maxdepth: 4
+
+   compatibility
+   getting_started
+   cloudant
+
+* :ref:`genindex`
+
