--- docs/indexes.rst
@@ -1,7 +1,0 @@
-indexes
-=======
-
-.. automodule:: cloudant.indexes
-    :members:
-    :undoc-members:
-    :show-inheritance:
